fix: flag new parcel to update fields on claim
